Split session implementation by responsibility
